refactor: 💡 Combine rescale computation with slice insertion
Combine rescale computation with slice insertion into volume.

Closes: #41
